telegram_weather_client/pyproject.toml
[tool.poetry]
name = "telegram_weather_client"
version = "0.1.0"
description = ""
authors = ["Max <brunomax23@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.26.0"
python-dotenv = "^1.0.1"
telethon = "^1.34.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
